[Gold I] Title: 비밀번호, Time: 128 ms, Memory: 16000 KB -BaekjoonHub

+++ 백준/Gold/2464. 비밀번호/비밀번호.java
@@ -28,7 +28,7 @@ public class Main {
 
 	static long nearestHigher(long A, long one, long mask, int oneCnt) {
 		long x = A ^ mask;                 // (01)->(10)로 스왑 (호출 조건이 01일 때만!)
-		long idx = 0;
+		int idx = 0;
 		while (((1L << idx) & mask) == 0) {
 			if (oneCnt > 0) {
 				oneCnt--;
